+++ src/main/php/predicate/Equals.php
@@ -14,7 +14,7 @@ use function bovigo\assert\export;
 /**
  * Predicate to test that something is equal.
  */
-class Equals extends Predicate
+class Equals extends Predicate implements Delta
 {
     /**
      * the expected value
@@ -34,6 +34,7 @@ class Equals extends Predicate
     /**
      * constructor
      *
+     * @deprecated  param  $delta is deprecated, use withDelta() on instance instead, will be removed with 7.0
      * @param  mixed  $expected  value to which test values must be equal
      * @param  float  $delta     allowed numerical distance between two values to consider them equal
      */
@@ -41,6 +42,18 @@ class Equals extends Predicate
     {
         $this->expected = $expected;
         $this->delta    = $delta;
+    }
+
+    /**
+     * sets delta which is allowed between expected and actual value
+     *
+     * @param   float  $delta
+     * @return  Predicate
+     */
+    public function withDelta(float $delta): Predicate
+    {
+        $this->delta = $delta;
+        return $this;
     }
 
     /**