+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## 6.0.0 (2020-02-??)
 
+### BC breaks
+
 * Raised minimum required PHP version to 7.3.0
 * Upgraded PHPUnit compatibility layer to PHPUnit 9.0
+* Deprecated optional parameter `$delta` in `bovigo\assert\predicate\equals()` and `bovigo\assert\predicate\isNotEqualTo()`, use new method `withDelta()` on returned instance
 
 ## 5.1.1 (2019-12-10)